Use world to store scheduler commands
